xiaohuihui-dify-tech-article/scripts/requirements.txt
# 腾讯云 COS SDK
cos-python-sdk-v5>=1.9.0

# 环境变量管理
python-dotenv>=0.19.0
